requirements.txt
discord.py==2.2.2
python-dotenv==0.21.0
emoji==1.7.0
requests==2.22.0
google-api-python-client==2.70.0
yt-dlp==2023.7.6
PyNaCl==1.3.0
isodate==0.6.1
urllib3==1.25.8
aiohttp==3.8.6
pytz==2024.1